updated netcdf_outputter to write all data if requested updated model.py with new interface of prepare_for_model_step

+++ py_gnome/gnome/netcdf_outputter.py
@@ -4,6 +4,7 @@ NetCDF outputter - follows the interface defined by gnome.Outputter for a NetCDF
 import copy
 import os
 from datetime import datetime
+from collections import OrderedDict
 
 import netCDF4 as nc
 import numpy as np
@@ -17,48 +18,72 @@ class NetCDFOutput(Outputter, serializable.Serializable):
     cf_attributes={'comment' : 'Particle output from the NOAA PyGnome model',
                    'source' : "PyGnome version x.x.x",
                    'references' : 'TBD',
-                   'CF:featureType' : "particle_trajectory" ,
+                   'feature_type' : "particle_trajectory" ,
                    'institution' : "NOAA Emergency Response Division",
                    'conventions' : "CF-1.6",
-                   'history' : "Evolved with discussion on CF-metadata listserve",
-                   #'title' : "Sample data/file for particle trajectory format",
                    }
     
-    data_vars= { 'longitude':{'dtype':np.float32,
-                              'units':'degrees_east',
-                              'long_name':'longitude of the particle'},
-                 
-                 'latitude': {'dtype':np.float32,
-                              'units':'degrees_north',
-                              'long_name':'latitude of the particle'},
-                 
-                 'depth':    {'dtype':np.float32,
-                              'units':'meters',
-                              'long_name':'particle depth below sea surface',
-                              'axis': 'z positive down'},
-                 
-                 'mass':     {'dtype':np.float32,
-                              'units':'grams'},
-                 
-                 'age':      {'dtype':np.int32,
-                              'units':'seconds',
-                              'long_name':'from age at time of release'},
-                 
-                 'status':   {'dtype':np.int8,
-                              'long_name': 'particle status flag',
-                              'valid_range': '0, 10',
-                              'flag_values': '2, 3, 7, 10',
-                              'flag_meanings': '2:in_water 3:on_land 7:off_maps 10:evaporated'
-                              },
-                 
-                 'id':       {'dtype':np.int8,
-                              'long_name': 'particle ID',
-                              'units' : '1'
-                              },
-                 }
+    """ let's keep order the same as original NetCDF """
+    data_vars = OrderedDict()
+    var = OrderedDict()
+    
+    # longitude
+    var['dtype'] = np.float32
+    var['long_name'] = 'longitude of the particle'
+    var['units'] = 'degrees_east'
+    data_vars['longitude'] = copy.deepcopy(var)
+    
+    # latitude
+    var['long_name'] = 'latitude of the particle'
+    var['units'] = 'degrees_north'
+    data_vars['latitude'] = copy.deepcopy(var)
+    
+    # latitude
+    var['long_name'] = 'particle depth below sea surface'
+    var['units'] = 'meters'
+    var['axis'] = 'z positive down'
+    data_vars['depth'] = copy.deepcopy(var)
+    
+    # mass
+    var.clear()
+    var['dtype'] = np.float32
+    var['units'] = 'grams'
+    data_vars['mass'] = copy.deepcopy(var)
+    
+    # age
+    var.clear()
+    var['dtype'] = np.int32
+    var['long_name'] = 'from age at time of release'
+    var['units'] = 'seconds'
+    data_vars['age'] = copy.deepcopy(var)
+    
+    # flag
+    var.clear()
+    var['dtype'] = np.int8
+    var['long_name'] = 'particle status flag'
+    var['valid_range'] = [0, 5]
+    var['flag_values'] = [1, 2, 3, 4],
+    var['flag_meanings'] = 'on_land off_maps evaporated below_surface'
+    data_vars['flag'] = copy.deepcopy(var)
+    
+    # status
+    var['long_name'] = 'particle status flag'
+    var['valid_range'] = [0, 10]
+    var['flag_values'] = [2, 3, 7, 10],
+    var['flag_meanings'] = '2:in_water 3:on_land 7:off_maps 10:evaporated'
+    data_vars['status'] = copy.deepcopy(var)
+    
+    # id
+    var.clear()
+    var['dtype'] = np.int8
+    var['long_name'] = 'particle ID'
+    var['units'] = '1'
+    data_vars['id'] = copy.deepcopy(var)
 
-
-    def __init__(self, netcdf_filename, cache=None, write_alldata=False):
+    # This is data that has already been written in standard format
+    standard_data = ['positions','current_time_stamp','status_codes','spill_num','age','mass','flag']
+    
+    def __init__(self, netcdf_filename, cache=None, all_data=False, id=None):
         """
         should netcdf_filename be overwritten of if it already exists?
         """
@@ -66,41 +91,47 @@ class NetCDFOutput(Outputter, serializable.Serializable):
         self.cache = cache
         self._uncertain = False
         self._u_netcdf_filename = None
-        self.write_alldata = write_alldata
+        self.all_data = all_data
+        self.arr_types = None   # this is only updated in prepare_for_model_run if all_data is True
         #self.del_on_rewind = del_on_rewind
-        
-        if os.path.exists(netcdf_filename):
-            raise ValueError("{0} file exists. Enter a filename that does not exist in which to save data.".format(netcdf_filename))
         
         if os.path.isdir(netcdf_filename):
             raise ValueError("netcdf_filename must be a file not a directory.")
         
+        if os.path.exists(netcdf_filename):
+            raise ValueError("{0} file exists. Enter a filename that does not exist in which to save data.".format(netcdf_filename))
+        
         if not os.path.exists( os.path.realpath(os.path.dirname(netcdf_filename))):
             raise ValueError("{0} does not appear to be a valid path".format(os.path.dirname(netcdf_filename)))
         
+        self._gnome_id = gnome.GnomeId(id)
     
-    def prepare_for_model_run(self, cache=None, **kwargs):
+    @property
+    def id(self):
+        return self._gnome_id.id
+    
+    def prepare_for_model_run(self, cache=None, model_start_time=None, num_time_steps=None, uncertain=False, spills=None, **kwargs):
         """ 
         Write global attributes
         
         :param cache=None: Sets the cache object to be used for the data.
                            If None, it will use the one already set up.
         
-        Since this takes more than standard 'cache' argument, those are in kwargs.
-        These are required arguments, so they do not contain defaults.
+        This takes more than standard 'cache' argument. These are required arguments - they contain None for defaults
+        because XXX
         :model_start_time: 
         :num_time_steps:
         
-        Does not take anyother input arguments; however, to keep the interface the same for all outputters,
+        Does not take any other input arguments; however, to keep the interface the same for all outputters,
         define **kwargs for now.
         """
         if cache is not None:
             self.cache = cache
         
-        model_start_time = kwargs.pop('model_start_time')
-        num_time_steps = kwargs.pop('num_time_steps')
+        if model_start_time is None or num_time_steps is None:
+            raise TypeError("model_start_time and num_time_steps cannot be NoneType")
         
-        self._uncertain = kwargs.pop('uncertain',self._uncertain)
+        self._uncertain = uncertain
         
         if self._uncertain:
             name, ext = os.path.splitext(self.netcdf_filename)
@@ -112,12 +143,13 @@ class NetCDFOutput(Outputter, serializable.Serializable):
         for file_ in filenames:
             with nc.Dataset(file_, 'w', format='NETCDF4') as rootgrp:
                 """ Global variables """
-                rootgrp.convention = self.cf_attributes['conventions']
-                rootgrp.institution = self.cf_attributes['institution']
-                rootgrp.source = self.cf_attributes['source']
-                rootgrp.history = self.cf_attributes['history']
                 rootgrp.comment = self.cf_attributes['comment']
                 rootgrp.creation_date = time_utils.round_time(datetime.now(),roundTo=1).isoformat().replace('T',' ')
+                rootgrp.source = self.cf_attributes['source']
+                rootgrp.references = self.cf_attributes['references']
+                rootgrp.feature_type = self.cf_attributes['feature_type']
+                rootgrp.institution = self.cf_attributes['institution']
+                rootgrp.convention = self.cf_attributes['conventions']
                 
                 """ Dimensions """
                 rootgrp.createDimension('time', num_time_steps)
@@ -126,6 +158,8 @@ class NetCDFOutput(Outputter, serializable.Serializable):
                 """ Variables """
                 time_ = rootgrp.createVariable('time', np.double, ('time',))
                 time_.units = 'seconds since {0}'.format(model_start_time.isoformat().replace('T',' '))
+                time_.long_name = 'time'
+                time_.standard_name = 'time'
                 time_.calendar = 'gregorian'
                 time_.comment = 'unspecfied time zone'
                 
@@ -139,9 +173,24 @@ class NetCDFOutput(Outputter, serializable.Serializable):
                     # iterate over remaining attributes
                     [setattr(var,key2,val2) for key2,val2 in val.iteritems() if key2 != 'dtype']
                     
-                #if self.write_alldata:
-                    # write all data
-            
+                """ End standard data. Next create variables for remaining arrays if all_data is True """ 
+                if self.all_data:
+                    rootgrp.createDimension('world_point', 3)
+                    self.arr_types = dict()
+                    for spill in spills:
+                        at = spill.array_types
+                        [self.arr_types.update({key:atype}) for key,atype in at.iteritems() if key not in self.arr_types and key not in self.standard_data]
+                    
+                    # create variables
+                    for key,val in self.arr_types.iteritems():
+                        if len(val.shape) == 0:
+                            rootgrp.createVariable(key, val.dtype,('data'))    
+                        elif val.shape[0] == 3:
+                            rootgrp.createVariable(key, val.dtype,('data','world_point'))
+                        else:
+                            raise ValueError("{0} has an undefined dimension: {1}".format(key,val.shape))
+                        
+    
     def write_output(self, step_num):
         """ write output at the end of the step"""
         if self.cache is None:
@@ -171,13 +220,24 @@ class NetCDFOutput(Outputter, serializable.Serializable):
                 rootgrp.variables['status'][ixs:ixe] = sc['status_codes'][:]
                 rootgrp.variables['id'][ixs:ixe] = sc['spill_num'][:]
                 
-            return {'step_num': step_num,
-                    'netcdf_filename': (self.netcdf_filename, self._u_netcdf_filename),
-                    'time_stamp': time_stamp}
+                # write remaining data
+                if self.all_data:
+                    for key, val in self.arr_types.iteritems():
+                        if len(val.shape) == 0:
+                            rootgrp.variables[key][ixs:ixe] = sc[key]
+                        else:
+                            rootgrp.variables[key][ixs:ixe,:] = sc[key]
+                    
+                
+        return {'step_num': step_num,
+                'netcdf_filename': (self.netcdf_filename, self._u_netcdf_filename),
+                'time_stamp': time_stamp}
             
     
     def rewind(self):
         """ if rewound, delete the file and start over? """
-        os.remove(self.netcdf_filename)
-        if self._u_netcdf_filename is not None:
+        if os.path.exists(self.netcdf_filename):
+            os.remove(self.netcdf_filename)
+            
+        if self._u_netcdf_filename is not None and os.path.exists(self._u_netcdf_filename):
             os.remove(self._u_netcdf_filename)
